05_frontend_clients/agrisensa_streamlit_dashboard/requirements.txt
streamlit
pandas
plotly
folium
streamlit-folium
requests
scikit-learn>=1.3.0
numpy
joblib
lightgbm
shap
opencv-python-headless
beautifulsoup4
lxml
google-generativeai
matplotlib
scipy
xlsxwriter
openpyxl
qrcode
